+++ CHANGELOG.md
@@ -3,12 +3,16 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+None
+
+## [0.2.0] - 2016-05-03
 #### Added
 - Transformer and Transformation classes.
 - Added APIFrameworkJSONRendererAppendTransformations delegate
 - Added phpunit to composer require-dev
 - Added unit tests for Transformation code
 - Controller names are based on full page path (#5)
+- Using PSR-4 folder structure for controllers. Controllers must have a namespace. (#7)
 
 #### Fixed
 - Checking that controller actually exists before trying to include it (#6)
@@ -33,5 +37,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Initial release
 - Added Symphony PDO as requirement
 
-[Unreleased]: https://github.com/pointybeard/api_framework/compare/v0.1.1...master
+[Unreleased]: https://github.com/pointybeard/api_framework/compare/v0.2.0...integration
+[0.1.2]: https://github.com/pointybeard/api_framework/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/pointybeard/api_framework/compare/v0.1.0...v0.1.1